requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==2.4.1
async-timeout==5.0.1
attrs==24.2.0
blinker==1.9.0
category-encoders==2.6.4
certifi==2024.8.30
charset-normalizer==3.4.0
choreographer==0.99.6
click==8.1.7
cloudpickle==3.1.0
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
cython==3.0.11
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
dependency-injector==4.42.0
deprecation==2.1.0
dnspython==2.7.0
email-validator==2.2.0
executing==2.1.0
fastapi==0.115.3
fastapi-cli==0.0.5
fastjsonschema==2.20.0
flask==3.0.3
fonttools==4.55.0
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
imbalanced-learn==0.12.4
importlib-metadata==8.5.0
ipython==8.29.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.4
joblib==1.3.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-core==5.7.2
jupyterlab-widgets==3.0.13
kaleido==0.4.1
kiwisolver==1.4.7
lightgbm==4.5.0
llvmlite==0.43.0
loguru==0.7.2
mako==1.3.6
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
nbformat==5.10.4
nest-asyncio==1.6.0
numba==0.60.0
numpy==1.26.4
orjson==3.10.11
packaging==24.2
pandas==2.1.4
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.0.0
plotly==5.24.1
plotly-resampler==0.10.0
pmdarima==2.0.4
prompt-toolkit==3.0.48
psutil==6.1.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.3
pycaret==3.3.2
pydantic==2.9.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
pyyaml==6.0.2
requests==2.32.3
retrying==1.3.4
rich==13.9.3
scikit-learn==1.4.2
scikit-plot==0.3.7
scipy==1.11.4
setuptools==75.5.0
shellingham==1.5.4
six==1.16.0
sktime==0.26.0
sniffio==1.3.1
sqlalchemy==2.0.36
starlette==0.41.0
statsmodels==0.14.4
tbats==1.1.3
tenacity==9.0.0
threadpoolctl==3.5.0
tqdm==4.67.0
traitlets==5.14.3
tsdownsample==0.1.3
typer==0.12.5
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
watchfiles==0.24.0
websockets==13.1
werkzeug==3.0.6
yellowbrick==1.5
zipp==3.21.0
grpcio==1.68.0
grpcio-tools==1.68.0